Dirty fix for surreal db Thing struct deserialization issue
This commit fixes the issue by using a Investment2 struct where the data type of id is string instead of Option<Thing> as a dirty workaround.

A fix for this has apprantly made https://github.com/surrealdb/surrealdb/pull/1659 so will need to fix this after the next surrealdb release.