Fix add investment in yew
This commit fixes the code to add investments by utilising the Investment2 struct as well as the API POST /inv.

In the POST request, the data of Investment2 is passed as json via the body by setting the header for json.

Note: this is just a basic mapping of the API, most of the fields are yet to be mapped.