Change the logic of renew investment
Add renew_investment callback.

Instead of editing the old inv to add the link to the renewed inv, we now add the link to the old inv instead and set the status of the renewed inv to renewed, which makes more sense.

This will also help in retrival of the data.